# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "postway API",
    "version": "1.0.0",
    "description": "Production-grade serverless REST API with TypeScript and Python Lambda support",
    "contact": {
      "name": "API Support",
      "email": "support@postway.services"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api-staging.postway.services",
      "description": "Staging environment"
    },
    {
      "url": "https://api.postway.services",
      "description": "Production environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "WorkOS JWT token obtained from authentication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Error message"
              },
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              }
            }
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Authentication required",
                "code": "UNAUTHORIZED"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Internal server error",
                "code": "INTERNAL_ERROR"
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
    "/v1/media/images": {
      "get": {
        "summary": "List user images",
        "description": "Retrieves a list of images uploaded by the authenticated user",
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter images by category"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 1000
            },
            "description": "Maximum number of images to return"
          },
          {
            "in": "query",
            "name": "continuationToken",
            "schema": {
              "type": "string"
            },
            "description": "Token for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "images": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "key": {
                                "type": "string"
                              },
                              "url": {
                                "type": "string"
                              },
                              "size": {
                                "type": "number"
                              },
                              "lastModified": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "category": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "continuationToken": {
                          "type": "string",
                          "description": "Token for retrieving next page"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/media/upload-image-direct": {
      "post": {
        "summary": "Upload an image directly to S3",
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "filename",
                  "contentType",
                  "imageData"
                ],
                "properties": {
                  "filename": {
                    "type": "string",
                    "description": "Original filename",
                    "example": "profile-photo.jpg"
                  },
                  "contentType": {
                    "type": "string",
                    "description": "MIME type of the image",
                    "example": "image/jpeg"
                  },
                  "imageData": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "description": "Category for organizing images",
                    "example": "profile"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "description": "S3 key of the uploaded image"
                        },
                        "url": {
                          "type": "string",
                          "description": "Public URL of the uploaded image"
                        },
                        "size": {
                          "type": "number",
                          "description": "Size of the uploaded image in bytes"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/media/upload-image": {
      "post": {
        "summary": "Generate presigned URL for image upload",
        "description": "Creates a presigned S3 URL for uploading user images",
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "filename",
                  "contentType"
                ],
                "properties": {
                  "filename": {
                    "type": "string",
                    "description": "Original filename of the image",
                    "example": "profile-photo.jpg"
                  },
                  "contentType": {
                    "type": "string",
                    "description": "MIME type of the image",
                    "enum": [
                      "image/jpeg",
                      "image/png",
                      "image/gif",
                      "image/webp"
                    ],
                    "example": "image/jpeg"
                  },
                  "category": {
                    "type": "string",
                    "description": "Category of the image (profile, document, etc.)",
                    "example": "profile"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned upload URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "uploadUrl": {
                          "type": "string",
                          "description": "Presigned URL for uploading the image"
                        },
                        "imageKey": {
                          "type": "string",
                          "description": "S3 key where the image will be stored"
                        },
                        "expiresIn": {
                          "type": "number",
                          "description": "URL expiry time in seconds"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "schema": {
              "$ref": "#/definitions/StandardErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/StandardErrorResponse"
            }
          }
        }
      }
    },
    "/v1/test/api-key": {
      "get": {
        "summary": "Test API key authentication",
        "description": "Tests the withApiKey middleware",
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "API key valid"
          },
          "401": {
            "description": "Invalid or missing API key"
          }
        }
      }
    },
    "/v1/test/webhook": {
      "post": {
        "summary": "Test webhook signature validation",
        "description": "Tests the withWebhookSignature middleware",
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Webhook-Signature",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC signature of the request body"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook signature valid"
          },
          "401": {
            "description": "Invalid signature"
          }
        }
      }
    },
    "/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user profile",
        "description": "Returns the authenticated user's complete profile including user and profile data",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object"
                        },
                        "profile": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "description": "Updates the authenticated user's profile. Only sends fields that need to be updated.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "object",
                    "properties": {
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "defaultTimezone": {
                        "type": "string"
                      }
                    }
                  },
                  "profile": {
                    "type": "object",
                    "properties": {
                      "preferredName": {
                        "type": "string"
                      },
                      "pronouns": {
                        "type": "string"
                      },
                      "location": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      },
                      "photoUrl": {
                        "type": "string"
                      },
                      "gender": {
                        "type": "string"
                      },
                      "lgbtq": {
                        "type": "boolean"
                      },
                      "ethnicity": {
                        "type": "string"
                      },
                      "languages": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "onboardingCompleted": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully"
          },
          "400": {
            "description": "Bad request - no fields to update"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/v1/webhooks/workos": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "WorkOS webhook handler",
        "description": "Handles WorkOS webhook events for user and organization lifecycle management.\nVerifies webhook signature and processes events idempotently.\n\n**Supported Events:**\n- `user.created` - Creates new user and auth identity\n- `user.updated` - Updates existing user data\n- `user.deleted` - Removes user and auth identity\n- `organization.created` - Creates new organization\n- `organization.updated` - Updates organization data\n- `organization.deleted` - Removes organization\n\n**Security:** Requires valid WorkOS webhook signature in headers.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Unique event ID",
                    "example": "evt_01H1234567890ABCDEFGHIJK"
                  },
                  "event": {
                    "type": "string",
                    "description": "Event type",
                    "enum": [
                      "user.created",
                      "user.updated",
                      "user.deleted",
                      "organization.created",
                      "organization.updated",
                      "organization.deleted"
                    ],
                    "example": "user.created"
                  },
                  "data": {
                    "type": "object",
                    "description": "Event payload (varies by event type)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "processed",
                            "already_processed"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  }
}